(PC-17817) chore(search): move sections to components folder
