+++ src/features/search/components/sections/tests/Location.native.test.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 import { LocationType } from 'features/search/enums'
 import { RadioButtonLocation } from 'features/search/pages/LocationModal'
 import { initialSearchState } from 'features/search/pages/reducer'
-import { Location } from 'features/search/sections/Location'
+import { Location } from 'features/search/components/sections/Location'
 import { GeoCoordinates } from 'libs/geolocation'
 import { render, fireEvent, superFlushWithAct } from 'tests/utils'
 import * as useModalAPI from 'ui/components/modals/useModal'
