+++ src/features/search/pages/DatesHoursModal.tsx
@@ -17,7 +17,7 @@ import { DATE_FILTER_OPTIONS } from 'features/search/enums'
 import { datesHoursSchema } from 'features/search/pages/schema/datesHoursSchema'
 import { useSearch } from 'features/search/pages/SearchWrapper'
 import { useGetFullscreenModalSliderLength } from 'features/search/pages/useGetFullscreenModalSliderLength'
-import { SectionTitle } from 'features/search/sections/titles'
+import { SectionTitle } from 'features/search/components/sections/titles'
 import { SearchState, SearchView } from 'features/search/types'
 import { useLogFilterOnce } from 'features/search/utils/useLogFilterOnce'
 import { formatToCompleteFrenchDate } from 'libs/parsers'
