+++ src/features/search/pages/LocationModal.tsx
@@ -18,7 +18,7 @@ import { useSearch } from 'features/search/pages/SearchWrapper'
 import { SuggestedPlaces } from 'features/search/pages/SuggestedPlaces'
 import { useGetFullscreenModalSliderLength } from 'features/search/pages/useGetFullscreenModalSliderLength'
 import { useSetFocusWithCondition } from 'features/search/pages/useSetFocusWithCondition'
-import { SectionTitle } from 'features/search/sections/titles'
+import { SectionTitle } from 'features/search/components/sections/titles'
 import { SearchState } from 'features/search/types'
 import { useLogFilterOnce } from 'features/search/utils/useLogFilterOnce'
 import { analytics } from 'libs/firebase/analytics'
