+++ src/features/search/pages/OfferTypeModal.tsx
@@ -14,8 +14,8 @@ import { SearchCustomModalHeader } from 'features/search/components/SearchCustom
 import { SearchFixedModalBottom } from 'features/search/components/SearchFixedModalBottom'
 import { initialSearchState } from 'features/search/pages/reducer'
 import { useSearch } from 'features/search/pages/SearchWrapper'
-import { OFFER_TYPES } from 'features/search/sections/OfferType'
-import { SectionTitle } from 'features/search/sections/titles'
+import { OFFER_TYPES } from 'features/search/components/sections/OfferType'
+import { SectionTitle } from 'features/search/components/sections/titles'
 import { SearchState, SearchView } from 'features/search/types'
 import { useLogFilterOnce } from 'features/search/utils/useLogFilterOnce'
 import { Form } from 'ui/components/Form'
