+++ src/features/search/pages/SearchFilter.tsx
@@ -10,7 +10,7 @@ import { LocationType } from 'features/search/enums'
 import { initialSearchState } from 'features/search/pages/reducer'
 import { MAX_RADIUS } from 'features/search/pages/reducer.helpers'
 import { useSearch } from 'features/search/pages/SearchWrapper'
-import Section from 'features/search/sections'
+import Section from 'features/search/components/sections'
 import { SearchView } from 'features/search/types'
 import { analytics } from 'libs/firebase/analytics'
 import { useGeolocation } from 'libs/geolocation'
