+++ src/features/search/pages/__tests__/DatesHoursModal.native.test.tsx
@@ -7,7 +7,7 @@ import { navigate } from '__mocks__/@react-navigation/native'
 import { DATE_FILTER_OPTIONS } from 'features/search/enums'
 import { DatesHoursModal } from 'features/search/pages/DatesHoursModal'
 import { initialSearchState } from 'features/search/pages/reducer'
-import { SectionTitle } from 'features/search/sections/titles'
+import { SectionTitle } from 'features/search/components/sections/titles'
 import { SearchView } from 'features/search/types'
 import { analytics } from 'libs/firebase/analytics'
 import { act, fireEvent, render, superFlushWithAct } from 'tests/utils'
