+++ src/features/search/pages/__tests__/OfferTypeModal.native.test.tsx
@@ -7,8 +7,8 @@ import { useUserProfileInfo } from 'features/profile/api'
 import { OfferType } from 'features/search/enums'
 import { OfferTypeModal } from 'features/search/pages/OfferTypeModal'
 import { initialSearchState } from 'features/search/pages/reducer'
-import { OFFER_TYPES } from 'features/search/sections/OfferType'
-import { SectionTitle } from 'features/search/sections/titles'
+import { OFFER_TYPES } from 'features/search/components/sections/OfferType'
+import { SectionTitle } from 'features/search/components/sections/titles'
 import { OfferTypes, SearchView } from 'features/search/types'
 import { analytics } from 'libs/firebase/analytics'
 import { fireEvent, render, act } from 'tests/utils'
