+++ src/features/search/pages/__tests__/SearchPrice.native.test.tsx
@@ -7,7 +7,7 @@ import { useUserProfileInfo } from 'features/profile/api'
 import { initialSearchState } from 'features/search/pages/reducer'
 import { MAX_PRICE } from 'features/search/pages/reducer.helpers'
 import { SearchPrice } from 'features/search/pages/SearchPrice'
-import { SectionTitle } from 'features/search/sections/titles'
+import { SectionTitle } from 'features/search/components/sections/titles'
 import { SearchView } from 'features/search/types'
 import { analytics } from 'libs/firebase/analytics'
 import { fireEvent, render, act, superFlushWithAct } from 'tests/utils'
